nexuskit/requirements.txt
fastapi
uvicorn
python-multipart
yt-dlp
PyPDF2
PyMuPDF
reportlab
Pillow
